+++ OUDS/Core/Themes/Wireframe/Sources/Values/SemanticTokens/WireframeTheme+FontCompositeSemanticTokens.swift
@@ -55,7 +55,7 @@ extension WireframeThemeFontSemanticTokensProvider: FontCompositeSemanticTokens 
     // MARK: - Semantic tokens - Typography - Composites - Label
 
     @objc public final var labelDefaultXLarge: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.regular300) }
-    @objc public final var labelDefaultLarge: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.bold250) }
+    @objc public final var labelDefaultLarge: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.regular250) }
     @objc public final var labelDefaultMedium: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.regular200) }
     @objc public final var labelDefaultSmall: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.regular150) }
     @objc public final var labelModerateXLarge: MultipleFontCompositeRawTokens { MultipleFontCompositeRawTokens(FontRawTokens.medium300) }